Consolidate hook watch path coverage
- Verify watch candidates across host, project, and worktree scopes
- Keep the coverage with the consolidated config-path test suite
